feat: Add support for submitting inventory in standalone mode
Two new configuration parameters specifies the inventory executable
(`API.InventoryExecutable`) and the interval between each execution
(`API.InventoryInterval`). The client keeps track of the checksum and will
not submit inventory as long as the attributes are the same.

The behavior when dbus API is enabled is unaltered, it is assumed that
the dbus Mender client will be responsible for inventory.

Changelog: Commit
Ticket: ALV-76
Signed-off-by: Alf-Rune Siqveland <alf.rune@northern.tech>